http://issues.civicrm.org/jira/browse/CRM-9582
This profile should contain the email address field only and should not be editable by CiviCRM end users. The profile should be set to reserved. As set of insert statements should be provided which create the data.

The tables in question are
•	civicrm_uf_group
•	civicrm_uf_field

The Email profile field (uf_field entry) should have location_type_id = NULL (this means Primary email, use default location type), and is_reserved = 1 (this will prevent admins from deleting or disabling this field in the reserved profile).

Signed-off-by: Mark Muhleder <mark@analogweb.co.uk>